Extract a Square.js
